+++ CHANGELOG.md
@@ -5,5 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [1.0.0]
-- #655796: Implement helper for extensions rtrim, ltrim, trim, split, round, repeat, translate, and next_day
+Version [0.1.0]
+--------------
+- Implement helper for extensions 
+- Adding function extension rtrim
+- Adding function extension ltrim
+- Adding function extension trim
+- Adding function extension split
+- Adding function extension round
+- Adding function extension repeat
+- Adding function extension translate
+- Adding function extension and next_day